Moved WIP classes over to Enhancement folder for organization purpose
