+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Enhancement/Config/GamePadConfig.java
@@ -1,8 +1,8 @@
-package org.firstinspires.ftc.teamcode.Config;
+package org.firstinspires.ftc.teamcode.Enhancement.Config;
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.Gamepad;
-import org.firstinspires.ftc.teamcode.Robot;
+import org.firstinspires.ftc.teamcode.Enhancement.Robot;
 
 /**
  * GamePadConfig has the Config for the gamepad, which is used in robot.java.
