+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Enhancement/Subsystems/Control/Control.java
@@ -1,10 +1,10 @@
-package org.firstinspires.ftc.teamcode.Subsystems.Control;
+package org.firstinspires.ftc.teamcode.Enhancement.Subsystems.Control;
 
 
 import com.qualcomm.robotcore.hardware.DcMotorEx;
 import com.qualcomm.robotcore.hardware.HardwareMap;
 import com.qualcomm.robotcore.util.ElapsedTime;
-import org.firstinspires.ftc.teamcode.Subsystems.Subsystem;
+import org.firstinspires.ftc.teamcode.Enhancement.Subsystems.Subsystem;
 import org.firstinspires.ftc.teamcode.Util.QuickTelemetry;
 
 public class Control extends Subsystem {
