+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Enhancement/Subsystems/Vision/DetectMarker/MarkerLocation.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.Subsystems.Vision.DetectMarker;
+package org.firstinspires.ftc.teamcode.Enhancement.Subsystems.Vision.DetectMarker;
 
 /**
  * The Marker Location is {@link #SEARCHING} by default.
