+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Teleop/OmniDirectionalDrive.java
@@ -7,8 +7,8 @@ import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
 import org.firstinspires.ftc.robotcore.external.navigation.AxesOrder;
 import org.firstinspires.ftc.robotcore.external.navigation.AxesReference;
 import org.firstinspires.ftc.robotcore.external.navigation.Orientation;
-import org.firstinspires.ftc.teamcode.Robot;
-import org.firstinspires.ftc.teamcode.Subsystems.Drive.Drive;
+import org.firstinspires.ftc.teamcode.Enhancement.Robot;
+import org.firstinspires.ftc.teamcode.Enhancement.Subsystems.Drive.Drive;
 import org.firstinspires.ftc.teamcode.Util.QuickTelemetry;
 
 import java.io.IOException;
