+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Teleop/Teleop.java
@@ -3,7 +3,7 @@ package org.firstinspires.ftc.teamcode.Teleop;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.util.ElapsedTime;
-import org.firstinspires.ftc.teamcode.Robot;
+import org.firstinspires.ftc.teamcode.Enhancement.Robot;
 
 /**
  * Normal Teleop, not omnidirectional drive.
